refactor(narrative): move LlmProviderInterface to Domain/Narrative/Services
The LLM provider interface is a domain contract for the Narrative
bounded context. Moving it from app/Services/ to Domain enforces
proper DDD boundaries.
