+++ backend/app/Domain/Narrative/Actions/GenerateNarrativesForReport.php
@@ -14,7 +14,7 @@ use App\Domain\Report\Models\ReportDay;
 use App\Domain\Report\Models\ReportDayTask;
 use App\Domain\Report\Models\ReportTask;
 use App\Domain\Report\ValueObjects\Narrative;
-use App\Services\LLM\LlmProviderInterface;
+use App\Domain\Narrative\Services\LlmProviderInterface;
 use Illuminate\Support\Facades\Log;
 use Throwable;
