+++ backend/app/Services/LLM/ClaudeProvider.php
@@ -9,6 +9,7 @@ use App\Domain\Narrative\DTOs\DayFallbackRequest;
 use App\Domain\Narrative\DTOs\DayFallbackResponse;
 use App\Domain\Narrative\DTOs\TaskNarrativeRequest;
 use App\Domain\Narrative\DTOs\TaskNarrativeResponse;
+use App\Domain\Narrative\Services\LlmProviderInterface;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Log;
 use RuntimeException;
