+++ backend/app/Services/LLM/LlmManager.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Services\LLM;
 
+use App\Domain\Narrative\Services\LlmProviderInterface;
 use InvalidArgumentException;
 
 class LlmManager
